+++ CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - feat: showcase usage of complete lb to return a response from app/webpage
 - feat: showcase `authenticate-cognito` with a working app/webpage
 
+## [1.1.3] - 2023-12-12
+- feat: showcase `authenticate-oidc` default action
+
 ## [1.1.2] - 2023-12-08
 - feat: showcase `authenticate-cognito` type listener
 - fix: listener should have one action error resulting from misconfiguration
@@ -95,8 +98,9 @@ Feature: Added listener certificate
 ### Changes
 - Initial commit
 
-[Unreleased]: https://github.com/boldlink/terraform-aws-lb/compare/1.1.2...HEAD
+[Unreleased]: https://github.com/boldlink/terraform-aws-lb/compare/1.1.3...HEAD
 
+[1.1.3]: https://github.com/boldlink/terraform-aws-lb/releases/tag/1.1.3
 [1.1.2]: https://github.com/boldlink/terraform-aws-lb/releases/tag/1.1.2
 [1.1.1]: https://github.com/boldlink/terraform-aws-lb/releases/tag/1.1.1
 [1.1.0]: https://github.com/boldlink/terraform-aws-lb/releases/tag/1.1.0